Split spec also
